+++ trading-app/src/main/java/com/bugbycode/repository/user/impl/UserRepositoryImpl.java
@@ -88,7 +88,7 @@ public class UserRepositoryImpl implements UserRepository {
 
 	@Override
 	public void updateBinanceApiSecurity(String username,String binanceApiKey, String binanceSecretKey,int autoTrade,
-			int baseStepSize,int leverage,int positionValue, int cutLoss,int profit,int autoTradeType) {
+			int baseStepSize,int leverage,int positionValue, int cutLoss,double profit,int autoTradeType) {
 		Update update = new Update();
 		update.set("binanceApiKey", binanceApiKey);
 		update.set("binanceSecretKey", binanceSecretKey);
